Ticket #4 -Criar mecanismo para desempate, com base nas regras de 2015 #4
Correções contidas no ticket #4
