--- src/main/groovy/br/com/zg/trainning/model/business/Interface/Pontuacao.groovy
@@ -1,9 +1,0 @@
-package br.com.zg.trainning.model.business.Interface
-
-import br.com.zg.trainning.model.entity.Time
-/**
- * Created by diegoalves on 07/04/16.
- */
-interface Pontuacao {
-	int CalculaPontuacao(Time time)
-}
